% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{check_hazard}
\alias{check_hazard}
\title{Validate an hazard object
If validation fails, an exception will be raised.}
\usage{
check_hazard(hazard, initPop = NULL)
}
\arguments{
\item{hazard}{An S3 object of class "golden_hazard"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\value{
No return value, called for side effects.
}
\description{
Validate an hazard object
If validation fails, an exception will be raised.
}
\examples{
library(data.table)
N <- 100
dt <- data.table(a = runif(N, 0, 1), b = rep(0, N))
# Define a hazard function, which returns a vector of equal length uncertainties
test_hazard <- function(a) {
    ret <- (a < 0.5)
}
# Define a transition function, which sets all "b" columns affected by the hazard to 100
test_transition <- function() {
    return (100)
}
# Create an S3 golden_hazard
haz <- new_hazard(
              test_hazard,
              c("a"),
              new_transition(test_transition, c(), "b")
            )
# check_hazard() will not throw an exception
# as haz is a valid S3 golden_hazard
# and dt contains column "a"
check_hazard(haz, dt)
}
