% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_cohort}
\alias{create_cohort}
\title{Create a new cohort}
\usage{
create_cohort(demog, N)
}
\arguments{
\item{demog}{Demographic information containing columns AgeGrp/PopMale/PopFemale/PopTotal}

\item{N}{Size of the population to generate}
}
\value{
A sample population data.table, with columns male/age/bmi/death
}
\description{
Temporary testing method, probably replaced in future with R's simdata package or similar
}
\examples{
library(data.table)
demog <- data.table(
  AgeGrp = c(0, 1, 2, 3),
  PopMale = c(1000, 1100, 1050, 980),
  PopFemale = c(950, 1020, 1005, 970)
)
demog[, PopTotal := PopMale + PopFemale]
cohort <- create_cohort(demog, 100)
}
