% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden-package.R
\docType{package}
\name{golden}
\alias{golden-package}
\alias{golden}
\title{Golden: Framework for Patient-Level Microsimulation of Risk Factor Trajectories & Hazard-Based Events}
\description{
Fast, flexible, patient-level microsimulation. Time-stepped simulation with a C++ back-end from user-supplied initial population, trajectories, hazards, and corresponding event transitions.
User-defined aggregate time series histories are returned together with the final population. Designed for simulation of chronic diseases with continuous and evolving risk factors, but could easily be applied more generally.
}
\examples{
\dontrun{
# A full example can be found in the vignettes
}
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk} (\href{https://orcid.org/0000-0001-5825-9347}{ORCID})

Authors:
\itemize{
  \item Robert Chisholm \email{robert.chisholm@sheffield.ac.uk} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})
}

}
