% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{new_column}
\alias{new_column}
\title{Create a new golden_history_column}
\usage{
new_column(name, fn, args, filter_fn = NULL, filter_args = NULL)
}
\arguments{
\item{name}{Name of the column in the output data-table}

\item{fn}{Reduction function, which converts the input columns to a single value}

\item{args}{Names of columns and special variables to be passed to fn}

\item{filter_fn}{(Optional) Filter function, which returns a bool vector denoting which rows should be reduced}

\item{filter_args}{(Optional) Names of columns and special variables to be passed to filter_fn. Required if filter_fn is}
}
\value{
An object of class "golden_history_column"
}
\description{
Create a new golden_history_column
}
