% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{new_hazard}
\alias{new_hazard}
\title{Create a new hazard object}
\usage{
new_hazard(fn, args, transitions, freq = 1, first = 0, last = 2147483647)
}
\arguments{
\item{fn}{Function which calculates the hazard likelihood}

\item{args}{Character vector of parameter names expected by fn}

\item{transitions}{Transition object(s) to be applied where the hazard is successful}

\item{freq}{(Optional) The frequency of hazard execution}

\item{first}{(Optional) First step the hazard should be enabled}

\item{last}{(Optional) Last step the hazard should be enabled}
}
\value{
An object of class "eldoradosim_hazard"
}
\description{
Create a new hazard object
}
