% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{new_history}
\alias{new_history}
\title{Create a new eldoradosim_history}
\usage{
new_history(columns, frequency = 1)
}
\arguments{
\item{columns}{eldoradosim_history_column S3 object(s)}

\item{frequency}{The number of simulation steps per history collection.}
}
\value{
An object of class "eldoradosim_history"
}
\description{
Create a new eldoradosim_history
}
