% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{new_parameters}
\alias{new_parameters}
\title{Create a new golden_parameters}
\usage{
new_parameters(
  hazards = list(),
  trajectories = list(),
  steps,
  random_seed = 0,
  debug = TRUE,
  print_timing = TRUE,
  history = NULL
)
}
\arguments{
\item{hazards}{golden_hazard S3 object(s)}

\item{trajectories}{golden_trajectory S3 object(s)}

\item{steps}{Number of steps to run}

\item{random_seed}{Seed to be used for random generation. If set 0, current R random state will be used.}

\item{debug}{(TRUE/FALSE) flag indicating whether validation checks are enabled. These catch NaN, but reduce performance}

\item{print_timing}{(TRUE/FALSE) flag indicating whether a per-function timing report should be printed after the simulation, this will always be suppressed for fast (<= 1 second) simulations.}

\item{history}{golden_history S3 object representing the columns of data to be aggregated during simulation}
}
\value{
An object of class "golden_parameters"
}
\description{
Create a new golden_parameters
}
\examples{
# Define a hazard function, which returns a vector of equal length uncertainties
test_hazard <- function(a) {
    ret <- (a < 0.5)
}
# Define a transition function, which sets all "b" columns affected by the hazard to 100
test_transition <- function() {
    return (100)
}
# Create an S3 golden_hazard
haz <- new_hazard(
              test_hazard,
              c("a"),
              new_transition(test_transition, c(), "b")
            )
# Define a trajectory function, which adds 2 to all members of the input vector
test_trajectory <- function(a) {
    return (a + 2)
}
# Create an S3 golden_trajectory
trj <- new_trajectory(test_trajectory, c("b"), "b")
# Create an S3 golden_history, containing 1 golden_history_column
hist <- new_history(new_column("sum_a", sum, c("a")))
# Create an S3 golden_parameters
params <- new_parameters(
  hazards = haz,
  trajectories = trj,
  steps = 10,
  debug = FALSE,
  history = hist
)
}
