% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{new_trajectory}
\alias{new_trajectory}
\title{Create a new trajectory object}
\usage{
new_trajectory(fn, args, states, name = NULL)
}
\arguments{
\item{fn}{Function defining the trajectory function}

\item{args}{Character vector of parameter names expected by fn}

\item{states}{Name(s) of the column(s) where the result(s) of the trajectory function is to be stored}

\item{name}{(Optional) Name used in error messages and similar. Defaults to an automatic name}
}
\value{
An object of class "golden_trajectory"
}
\description{
Create a new trajectory object
}
\note{
If a list if passed to states, fn must return a list of equal length
}
\examples{
# Define a trajectory function, which adds 2 to all members of the input vector
test_trajectory <- function(a) {
    return (a + 2)
}
# Create an S3 golden_trajectory
trj <- new_trajectory(test_trajectory, c("b"), "b")
}
