% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{new_transition}
\alias{new_transition}
\title{Create a new transition object}
\usage{
new_transition(fn, args, states, name = NULL)
}
\arguments{
\item{fn}{Function defining the transition functions}

\item{args}{Character vector of parameter names expected by fn}

\item{states}{Name(s) of the column(s) where the result of the transition function is to be stored}

\item{name}{(Optional) Name used in error messages and similar. Defaults to an automatic name}
}
\value{
An object of class "golden_transition"
}
\description{
Create a new transition object
}
\examples{
# Define a transition function, which sets all columns affected by the hazard to 100
test_transition <- function() {
    return (100)
}
# Define an S3 golden_transition
trn <- new_transition(test_transition, c(), "b")
}
