% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{print.golden_hazard}
\alias{print.golden_hazard}
\title{Print the contents of a golden_hazard type S3 object}
\usage{
\method{print}{golden_hazard}(x, ..., indent = 0L)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}

\item{indent}{(Optional) The level the printing is indented, useful if nested within another S3 object}
}
\value{
No return value, called for side effects.
}
\description{
Print the contents of a golden_hazard type S3 object
}
\examples{
# Define a hazard function, which returns a vector of equal length uncertainties
test_hazard <- function(a) {
    ret <- (a < 0.5)
}
# Define a transition function, which sets all "b" columns affected by the hazard to 100
test_transition <- function() {
    return (100)
}
# Create an S3 golden_hazard
haz <- new_hazard(
              test_hazard,
              c("a"),
              new_transition(test_transition, c(), "b")
            )
print(haz)
}
