% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{print.golden_history}
\alias{print.golden_history}
\title{Print the contents of a golden_history type S3 object}
\usage{
\method{print}{golden_history}(x, ..., indent = 0L)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}

\item{indent}{(Optional) The level the printing is indented, useful if nested within another S3 object}
}
\value{
No return value, called for side effects.
}
\description{
Print the contents of a golden_history type S3 object
}
\examples{
# Create an S3 golden_history
hist <- new_history(new_column("sum_a", sum, c("a")))
print(hist)
}
