% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{print.golden_history_column}
\alias{print.golden_history_column}
\title{Print the contents of a golden_history_column type S3 object}
\usage{
\method{print}{golden_history_column}(x, ..., indent = 0L)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}

\item{indent}{(Optional) The level the printing is indented, useful if nested within another S3 object}
}
\description{
Print the contents of a golden_history_column type S3 object
}
