% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{print.golden_timing}
\alias{print.golden_timing}
\title{Print the contents of a golden_timing type S3 object}
\usage{
\method{print}{golden_timing}(x, ...)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}
}
\description{
Print the contents of a golden_timing type S3 object
}
