% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{print.golden_trajectory}
\alias{print.golden_trajectory}
\title{Print the contents of a golden_trajectory type S3 object}
\usage{
\method{print}{golden_trajectory}(x, ..., indent = 0L)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}

\item{indent}{(Optional) The level the printing is indented, useful if nested within another S3 object}
}
\value{
No return value, called for side effects.
}
\description{
Print the contents of a golden_trajectory type S3 object
}
\examples{
# Define a trajectory function, which adds 2 to all members of the input vector
test_trajectory <- function(a) {
    return (a + 2)
}
# Create an S3 golden_trajectory
trj <- new_trajectory(test_trajectory, c("b"), "b")
print(trj)
}
