% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_simulation}
\alias{run_simulation}
\title{Execute a patient trajectory simulation}
\usage{
run_simulation(initPop, parameters)
}
\arguments{
\item{initPop}{data.table containing initial population for simulation}

\item{parameters}{Simulation configuration}
}
\value{
An list containing final population, history and timing data.tables
}
\description{
Execute a patient trajectory simulation
}
\examples{
library(data.table)
N <- 100
dt <- data.table(a = runif(N, 0, 1), b = rep(0, N))
# Define a hazard function, which returns a vector of equal length uncertainties
test_hazard <- function(a) {
    ret <- (a < 0.5)
}
# Define a transition function, which sets all "b" columns affected by the hazard to 100
test_transition <- function() {
    return (100)
}
# Create an S3 golden_hazard
haz <- new_hazard(
              test_hazard,
              c("a"),
              new_transition(test_transition, c(), "b")
            )
# Define a trajectory function, which adds 2 to all members of the input vector
test_trajectory <- function(a) {
    return (a + 2)
}
# Define an S3 golden_trajectory
trj <- new_trajectory(test_trajectory, c("b"), "b")
# Create an S3 golden_history, containing 1 golden_history_column
hist <- new_history(new_column("sum_a", sum, c("a")))
# Define an S3 golden_parameters
params <- new_parameters(
  hazards = haz,
  trajectories = trj,
  steps = 10,
  debug = FALSE,
  history = hist
)
# Run the simulation to collect results
results <- run_simulation(dt, params)
}
