% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{check_column}
\alias{check_column}
\title{Validate an history column object}
\usage{
check_column(column, initPop = NULL)
}
\arguments{
\item{column}{An S3 object of class "golden_history_column"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\description{
Validate an history column object
}
