% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{check_hazard}
\alias{check_hazard}
\title{Validate an hazard object}
\usage{
check_hazard(hazard, initPop = NULL)
}
\arguments{
\item{hazard}{An S3 object of class "eldoradosim_hazard"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\description{
Validate an hazard object
}
