% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{check_history}
\alias{check_history}
\title{Validate an history object}
\usage{
check_history(history, initPop = NULL)
}
\arguments{
\item{history}{An S3 object of class "golden_history"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\description{
Validate an history object
}
