% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Validate the configuration passed to run_simulation()}
\usage{
check_parameters(parameters, initPop = NULL)
}
\arguments{
\item{parameters}{An eldoradosim_parameters S3 object to be validated}

\item{initPop}{data.frame which contains the columns required by parameters}
}
\description{
Validate the configuration passed to run_simulation()
}
