% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{check_trajectory}
\alias{check_trajectory}
\title{Validate an trajectory object}
\usage{
check_trajectory(trajectory, initPop = NULL)
}
\arguments{
\item{trajectory}{An S3 object of class "eldoradosim_trajectory"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\description{
Validate an trajectory object
}
