% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{check_transition}
\alias{check_transition}
\title{Validate an transition object}
\usage{
check_transition(transition, initPop = NULL)
}
\arguments{
\item{transition}{An S3 object of class "eldoradosim_transition"}

\item{initPop}{(Optional) data.table to check columns required by functions exist}
}
\description{
Validate an transition object
}
