% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_cohort}
\alias{create_cohort}
\title{Create a new cohort}
\usage{
create_cohort(demog, N)
}
\arguments{
\item{demog}{}

\item{N}{}
}
\value{
A sample population data.table
}
\description{
Temporary testing method, probably replaced in future with R's simdata package or similar
}
