% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globorisk_coefs}
\alias{globorisk_coefs}
\title{Example globorisk coefficients}
\format{
TODO
\describe{
  \item{foo}{bar}
}
}
\source{
TODO
}
\usage{
globorisk_coefs
}
\description{
details TODO
}
\keyword{datasets}
