% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifetable_data}
\alias{lifetable_data}
\title{Example life table data}
\format{
TODO
\describe{
  \item{foo}{bar}
}
}
\source{
TODO
}
\usage{
lifetable_data
}
\description{
details TODO
}
\keyword{datasets}
