% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{new_parameters}
\alias{new_parameters}
\title{Create a new golden_parameters}
\usage{
new_parameters(
  hazards = list(),
  trajectories = list(),
  steps,
  random_seed = 0,
  debug = TRUE,
  print_timing = TRUE,
  history = NULL
)
}
\arguments{
\item{hazards}{golden_hazard S3 object(s)}

\item{trajectories}{golden_trajectory S3 object(s)}

\item{steps}{Number of steps to run}

\item{random_seed}{Seed to be used for random generation. If set 0, current R random state will be used.}

\item{debug}{(TRUE/FALSE) flag indicating whether validation checks are enabled. These catch NaN, but reduce performance}

\item{print_timing}{(TRUE/FALSE) flag indicating whether a per-function timing report should be printed after the simulation, this will always be suppressed for fast (<= 1 second) simulations.}

\item{history}{golden_history S3 object representing the columns of data to be aggregated during simulation}
}
\value{
An object of class "golden_parameters"
}
\description{
Create a new golden_parameters
}
