% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{new_transition}
\alias{new_transition}
\title{Create a new transition object}
\usage{
new_transition(fn, args, state, name = NULL)
}
\arguments{
\item{fn}{Function defining the transition functions}

\item{args}{Character vector of parameter names expected by fn}

\item{state}{Name(s) of the column(s) where the result of the transition function is to be stored}

\item{name}{(Optional) Name used in error messages and similar. Defaults to an automatic name}
}
\value{
An object of class "golden_transition"
}
\description{
Create a new transition object
}
