% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_snapshot}
\alias{pop_snapshot}
\title{Example population structure}
\format{
TODO
\describe{
  \item{foo}{bar}
}
}
\source{
TODO
}
\usage{
pop_snapshot
}
\description{
details TODO
}
\keyword{datasets}
