% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{print.golden_parameters}
\alias{print.golden_parameters}
\title{Print the contents of a golden_parameters type S3 object}
\usage{
\method{print}{golden_parameters}(x, ..., indent = 0)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Not used. Included for S3 method compatibility.}

\item{indent}{(Optional) The level the printing is indented, useful if nested within another S3 object}
}
\description{
Print the contents of a golden_parameters type S3 object
}
