% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_simulation}
\alias{run_simulation}
\title{Execute a patient trajectory simulation}
\usage{
run_simulation(initPop, parameters)
}
\arguments{
\item{initPop}{data.table containing initial population for simulation}

\item{parameters}{Simulation configuration}
}
\value{
An list containing final population, history and timing data.tables
}
\description{
Execute a patient trajectory simulation
}
